refactor files - split spot and perps
